Crate libp2p_metrics

Auxiliary crate recording protocol and Swarm events and exposing them as metrics in the OpenMetrics format.

See examples directory for more.

Structs

Metrics

Set of Swarm and protocol metrics derived from emitted events.

Traits

Recorder

Recorder that can record Swarm and protocol events.